GMLAS: compiler warning fixes
git-svn-id: https://svn.osgeo.org/gdal/trunk@36209 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gmlas/ogrgmlaswriter.cpp
@@ -376,7 +376,7 @@ bool GMLASWriter::Write(GDALProgressFunc pfnProgress,
         }
         poOtherMetadataLayer->ResetReading();
 
-        for( size_t i = 1; i <= oMapToUri.size(); ++i )
+        for( int i = 1; i <= static_cast<int>(oMapToUri.size()); ++i )
         {
             if( oMapToUri.find(i) != oMapToUri.end() )
             {
